+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# DONE (2023-11-20)
+
+- [x] add queue
+- [x] testes unitarios com queue
+
+# DONE (2023-11-19)
+
+- [x] testes unitarios com stack
+- [x] add ct-cake to template
+- [x] add test suit to template
+
 # DONE (2023-11-18)
 
 - [x] add stack